[refactor] SignupScreen1 -> SignupScreen으로 이름 변경 #95
